pkvm: move source changes into patch phase

+++ tools/pkvm-aarch64/patches/overlay/scripts/create_hostimg.sh
@@ -3,22 +3,27 @@ set -e -o pipefail
 
 . "$BASE_DIR/scripts/img_util.sh"
 
-export PATH=$PATH:/usr/sbin
+ensure_root_namespace "$@"
+
+export PATH=../buildtools/usr/bin:../buildtools/usr/sbin:$PATH:/usr/sbin
 cd "$(dirname "$0")"
 
-QEMU_USER=$(which qemu-aarch64-static)
+QEMU_USER=$BASE_DIR/oss/ubuntu/usr/bin/qemu-aarch64-static
+QEMU_HOST="${QEMU_HOST:-$BASE_DIR/oss/ubuntu/usr/bin/qemu-system-aarch64}"
+QEMU_VIRTIO_ROM=$BASE_DIR/oss/ubuntu/usr/share/qemu/efi-virtio.rom
+CROSVM=$BASE_DIR/oss/ubuntu/usr/bin/crosvm
 CPUS=$(nproc)
 
 USERNAME=$1
 CURDIR=$PWD
-PKGLIST=$(cat package.list.22 | grep -v "\-dev")
+PKGLIST=$(cat package.list.22)
 EXTRA_PKGLIST=$(cat extra_package.list)
-OUTFILE=ubuntuguest.qcow2
-RAWFILE=ubuntuguest.raw
+OUTFILE=ubuntuhost.qcow2
+RAWFILE=ubuntuhost.raw
 IMAGESDIR=$BASE_DIR/images
-OUTDIR=$IMAGESDIR/guest
+OUTDIR=$IMAGESDIR/host
 UBUNTUTEMPLATE=$BASE_DIR/oss/ubuntu-template
-SIZE=10G
+SIZE=20G
 START_BYTES=1048576
 SKIP_CHROOT=0
 
@@ -67,9 +72,33 @@ parted -s "$RAWFILE" mklabel gpt mkpart primary ext4 1MiB 100%
 
 echo "Copying ubuntu from template.."
 mkdir -p tmp
-sudo tar -C "$UBUNTUTEMPLATE" --numeric-owner -cf - ./ | tar -C tmp --numeric-owner -xf -
-cp "$QEMU_USER" tmp/usr/bin
-mkdir -p tmp/etc/network
+tar -C "$UBUNTUTEMPLATE" --numeric-owner -cf - ./ | tar -C tmp --numeric-owner -xf -
+mkdir -p tmp/etc/network tmp/home/ubuntu
+
+if [ ! -f "$QEMU_USER" ] || [ ! -f "$QEMU_HOST" ]; then
+	if [ ! -f "$CROSVM" ]; then
+		echo "ERROR: can't find a VMM"
+		echo "ERROR: please run 'make target-qemu' or 'make target-crosvm"
+		exit 1
+	fi
+fi
+
+if [ -f "$QEMU_USER" ]; then
+	cp "$QEMU_USER" tmp/usr/bin
+fi
+
+if [ -f "$QEMU_HOST" ]; then
+	cp "$QEMU_HOST" tmp/usr/bin
+fi
+
+if [ -f "$QEMU_VIRTIO_ROM" ]; then
+	mkdir -p tmp/usr/share/qemu
+	install --mode=0644 "$QEMU_VIRTIO_ROM" tmp/usr/share/qemu
+fi
+
+if [ -f "$CROSVM" ]; then
+	cp "$CROSVM" tmp/usr/bin
+fi
 
 echo "Installing packages.."
 if mount --bind /dev tmp/dev && mount -t proc none tmp/proc; then
@@ -77,7 +106,6 @@ if mount --bind /dev tmp/dev && mount -t proc none tmp/proc; then
 	export DEBIAN_FRONTEND=noninteractive
 	run_chroot tmp /bin/bash -lc "apt-get update"
 	run_chroot tmp /bin/bash -lc "apt-get -y install $EXTRA_PKGLIST"
-	run_chroot tmp /bin/bash -lc "apt-get -y purge network-manager network-manager-gnome network-manager-pptp"
 	run_chroot tmp /bin/bash -lc "update-alternatives --set iptables /usr/sbin/iptables-legacy"
 	run_chroot tmp /bin/bash -lc "adduser --disabled-password --gecos '' ubuntu"
 	run_chroot tmp /bin/bash -lc "passwd -d ubuntu"
@@ -95,15 +123,15 @@ cat >> tmp/etc/network/interfaces << EOF
 auto lo
 iface lo inet loopback
 
-auto enp0s2
-iface enp0s2 inet static
-address 192.168.10.3
-gateway 192.168.10.1
+auto enp0s1
+iface enp0s1 inet static
+address 192.168.7.2
+gateway 192.168.7.1
 EOF
 
 cat >> tmp/etc/hosts << EOF
 127.0.0.1	localhost
-127.0.1.1	pkvm-guest
+127.0.1.1	pkvm-host
 
 ::1     ip6-localhost ip6-loopback
 fe00::0 ip6-localnet
@@ -113,7 +141,7 @@ ff02::2 ip6-allrouters
 
 EOF
 
-echo pkvm-guest > tmp/etc/hostname
+echo pkvm-host > tmp/etc/hostname
 if [ -f tmp/etc/systemd/resolved.conf ]; then
 	sed 's/#DNS=/DNS=8.8.8.8/' -i tmp/etc/systemd/resolved.conf
 fi
@@ -121,22 +149,18 @@ if [ -f tmp/etc/ssh/sshd_config ]; then
 	sed 's/#PermitEmptyPasswords no/PermitEmptyPasswords yes/' -i tmp/etc/ssh/sshd_config
 fi
 
-echo 'DefaultTimeoutStartSec=600s' >> tmp/etc/systemd/system.conf
-mkdir -p tmp/etc/systemd/system/ifupdown-pre.service.d
-cat >> tmp/etc/systemd/system/ifupdown-pre.service.d/override.conf << EOF
-[Service]
-ExecStart=
-ExecStart=/bin/sh -c 'if [ "$CONFIGURE_INTERFACES" != "no" ] && [ -n "$(ifquery --read-environment --list --exclude=lo)" ] && [ -x /bin/udevadm ]; then udevadm settle --timeout 300; fi'
-TimeoutStartSec=400
-EOF
+echo "Installing modules.."
+make -C"$CURDIR/../linux-host" CROSS_COMPILE=aarch64-linux-gnu- ARCH=arm64 INSTALL_MOD_STRIP=1 INSTALL_MOD_PATH="$CURDIR/tmp" modules_install
+make -C"$CURDIR/../linux-host" CROSS_COMPILE=aarch64-linux-gnu- ARCH=arm64 INSTALL_HDR_PATH="$CURDIR/tmp/usr" headers_install
 
-pwd_dir=$(pwd)
-INST_MOD_PATH="$(pwd)/tmp"
-INST_HDR_PATH="$(pwd)/tmp/usr"
-echo "Installing guest kernel modules.."
-make -C"$GUEST_KERNEL_DIR" CROSS_COMPILE=aarch64-linux-gnu- ARCH=arm64 INSTALL_MOD_STRIP=1 INSTALL_MOD_PATH="$INST_MOD_PATH" modules_install
-make -C"$GUEST_KERNEL_DIR" CROSS_COMPILE=aarch64-linux-gnu- ARCH=arm64 INSTALL_HDR_PATH="$INST_HDR_PATH" headers_install
-echo Done
+install --mode=0755 "$BASE_DIR/scripts/run-crosvm.sh" tmp/usr/bin
+if [ "$SKIP_CHROOT" = "0" ]; then
+	run_chroot tmp /bin/bash -lc "chown root:root /usr/bin/run-crosvm.sh"
+	install --mode=0755 "$BASE_DIR/scripts/run-crosvm.sh" tmp/home/ubuntu
+	run_chroot tmp /bin/bash -lc "chown ubuntu:ubuntu /home/ubuntu/run-crosvm.sh"
+else
+	chown 0:0 tmp/usr/bin/run-crosvm.sh || true
+fi
 
 if [ ! -d "$OUTDIR" ]; then
 	echo "Creating output dir.."
@@ -144,11 +168,9 @@ if [ ! -d "$OUTDIR" ]; then
 	chown -R "$USERNAME.$USERNAME" "$IMAGESDIR"
 fi
 
-cp -f "$GUEST_KERNEL_DIR/arch/arm64/boot/Image" "$OUTDIR"
-chown "$USERNAME.$USERNAME" "$OUTDIR/Image"
 do_unmount tmp/proc || true
 do_unmount tmp/dev || true
 finalize_raw_image tmp "$RAWFILE" "$OUTDIR/$OUTFILE" "$START_BYTES"
 rm -f "$RAWFILE"
-echo "Output saved at $OUTDIR"
+echo "Output saved at $OUTDIR/$OUTFILE"
 sync
